refactored landing page to live view
