+++ lib/mindwendel_web/controllers/static_page_controller.ex
@@ -1,27 +1,8 @@
 defmodule MindwendelWeb.StaticPageController do
   use MindwendelWeb, :controller
-  alias Mindwendel.Brainstormings
-  alias Mindwendel.Brainstormings.Brainstorming
   alias Mindwendel.FeatureFlag
 
   plug :put_root_layout, {MindwendelWeb.Layouts, :static_page}
-
-  def home(conn, _params) do
-    current_user =
-      conn
-      |> Mindwendel.Services.SessionService.get_current_user_id()
-      |> Mindwendel.Accounts.get_user()
-
-    form =
-      %Brainstorming{}
-      |> Brainstormings.change_brainstorming(%{})
-      |> Phoenix.Component.to_form()
-
-    render(conn, "home.html",
-      current_user: current_user,
-      form: form
-    )
-  end
 
   def legal(conn, _params) do
     if FeatureFlag.enabled?(:feature_privacy_imprint_enabled) do
