+++ lib/mindwendel_web/templates/layout/app_static.html.heex
@@ -1,4 +1,4 @@
-<main role="main" class="container" id="main-container">
+<main role="main" class="container d-flex h-100 bg-black" id="main-container">
   <%= if Phoenix.Flash.get(@flash, :info) do %>
     <p class="alert alert-info" role="alert">
       {Phoenix.Flash.get(@flash, :info)}
