+++ lib/mindwendel_web/templates/layout/static_page.html.heex
@@ -16,7 +16,7 @@
     </script>
   </head>
 
-  <body class="d-flex h-100 bg-black">
+  <body class="h-100 bg-black">
     {@inner_content}
   </body>
 </html>
